Force app to use WIFI to make requests
Added code that forces all sockets created in the app to send their
requests via WIFI. This is done to prevent LAN requests being made
through mobile data.
