+++ app/src/main/java/br/org/cesar/knot_setup_app/activity/login/LoginActivity.java
@@ -12,7 +12,7 @@ import android.widget.TextView;
 import br.org.cesar.knot_setup_app.R;
 import br.org.cesar.knot_setup_app.activity.login.LoginContract.Presenter;
 import br.org.cesar.knot_setup_app.activity.thingList.ThingActivity;
-import br.org.cesar.knot_setup_app.utils.Utils;
+import br.org.cesar.knot_setup_app.wrapper.NetworkWrapper;
 
 public class LoginActivity extends AppCompatActivity implements LoginContract.ViewModel {
         private Presenter mPresenter;
@@ -41,8 +41,8 @@ public class LoginActivity extends AppCompatActivity implements LoginContract.Vi
                     mPresenter.doLogin(passFld.getText().toString());
                 }
             });
-
-            if(Utils.isConnected(this)){
+            //TODO: replace
+            if(NetworkWrapper.isConnected(this)){
                 mPresenter.fillEmail();
             }
         }
