+++ app/src/main/java/br/org/cesar/knot_setup_app/activity/splash/SplashActivity.java
@@ -5,25 +5,30 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.Toast;
 
+import br.org.cesar.knot_setup_app.KnotSetupApplication;
 import br.org.cesar.knot_setup_app.R;
 import br.org.cesar.knot_setup_app.activity.login.LoginActivity;
 import br.org.cesar.knot_setup_app.activity.splash.SplashContract.ViewModel;
 import br.org.cesar.knot_setup_app.activity.splash.SplashContract.Presenter;
 import br.org.cesar.knot_setup_app.activity.thingList.ThingActivity;
-import br.org.cesar.knot_setup_app.utils.Utils;
+import br.org.cesar.knot_setup_app.wrapper.NetworkWrapper;
 
 public class SplashActivity extends AppCompatActivity implements  ViewModel{
 
     private Presenter mPresenter;
-
+    private NetworkWrapper networkWrapper;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
+
+        networkWrapper = KnotSetupApplication.getNetworkWrapper();
+        networkWrapper.forceSocketsThroughWifi();
+
         mPresenter = new SplashPresenter(this,this);
 
-        if(Utils.isConnected(this)){
+        if(NetworkWrapper.isConnected(this)){
             mPresenter.chooseActivity();
         }
     }
