--- app/src/main/java/br/org/cesar/knot_setup_app/utils/Utils.java
@@ -1,21 +1,0 @@
-package br.org.cesar.knot_setup_app.utils;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-
-public class Utils {
-
-    public static boolean isConnected(Context context) {
-        ConnectivityManager connectivityManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-
-        if (connectivityManager != null) {
-            NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
-
-            return activeNetworkInfo != null
-                    && activeNetworkInfo.isConnected();
-        }
-
-        return false;
-    }
-}
